added options.data - changed partials and helpers processing
options.data can be a string or a boolean or undefined

if options.data is a string, it should represent a path to a folder
containing data files (json/yaml)

if options.data is set to TRUE, the compiler will look for json/yaml
files sharing the template files name at the same level.

if options.data is undefined or FALSE, no data file will be processed

---

partials and helpers parameters should be strings or array of strings
it should be path to folders containing partial files or helper file

helper file should contain only the logic and not be wrapped in
Handlebars.registerHelper() as the plugin does it for you.